WIP: reorganize files
